Darken menu border color
The lightgray menu border color #ededed introduced in
commit f147e5a66f43657805542904b43e0b4e662ffe04 (by PR #2095)
is way too light for my display (and probably others).
Thus the menus have no clear border which looks odd and broken.

This commit simply darkens the menu border to #bdbdbd, which
coincides with the (current default) color of the menu separator.

Signed-off-by: Stephan Beyer <s-beyer@gmx.net>